+++ sentry/src/test/java/io/sentry/transport/AsyncHttpTransportTest.kt
@@ -15,8 +15,10 @@ import io.sentry.SentryEnvelopeHeader
 import io.sentry.SentryEnvelopeItem
 import io.sentry.SentryEvent
 import io.sentry.SentryOptions
+import io.sentry.SentryOptionsManipulator
 import io.sentry.Session
 import io.sentry.TypeCheckHint.SENTRY_TYPE_CHECK_HINT
+import io.sentry.clientreport.NoOpClientReportRecorder
 import io.sentry.dsnString
 import io.sentry.protocol.User
 import java.io.IOException
@@ -35,6 +37,7 @@ class AsyncHttpTransportTest {
             setSerializer(mock())
             setEnvelopeDiskCache(mock())
         }
+        var clientReportRecorder = NoOpClientReportRecorder()
 
         init {
             // this is an executor service running immediately in the current thread. Of course this defeats the
@@ -44,6 +47,7 @@ class AsyncHttpTransportTest {
         }
 
         fun getSUT(): AsyncHttpTransport {
+            SentryOptionsManipulator.setClientReportRecorder(sentryOptions, clientReportRecorder)
             return AsyncHttpTransport(executor, sentryOptions, rateLimiter, transportGate, connection)
         }
     }