+++ CHANGELOG.md
@@ -3,7 +3,9 @@
 ## Unreleased
 
 * Fix: Profiling rate decreased from 300hz to 100hz; fixed profiling traces folder creation on manual sdk init (#1997)
+* Feat: Add sentry-servlet-jakarta module (#1987)
 * Ref: Make options.printUncaughtStackTrace primitive type (#1995)
+* Feat: Add client reports (#1982)
 
 ## 6.0.0-alpha.5